requirements.txt
DateTime==6.0
folium==0.20.0
haversine==2.9.0
plotly==6.6.0
plotly-express==0.4.1
streamlit==1.55.0
streamlit-folium==0.27.
pillow==10.4.0
pandas==2.2.2
numpy==2.1.0